components: add ProjectCard
